fix: refactor code and change writing file technique

+++ packages/nyerati/functions/parsecoord.js
@@ -1,32 +1,21 @@
 const fs = require("fs");
 const path = require("path");
 
-const recordFolder = path.join(__dirname, "..", "record");
-const recordText = path.join(recordFolder, "coord.txt");
-const recordJSON = path.join(recordFolder, "convert.json");
+const {
+  config: {
+    constant: { recordFolder, recordText, recordJson },
+  },
+} = require("@nyerati/shared")(process);
 
-const parseCoordTxt = () =>
-  fs
+if (fs.existsSync(recordFolder) && fs.existsSync(recordText)) {
+  const data = fs
     .readFileSync(recordText, "utf8")
     .split(/\r\n|\r|\n/gm)
     .filter((d) => d !== "")
     .map(JSON.parse);
 
-if (fs.existsSync(recordFolder)) {
-  fs.open(recordText, "r", (err) => {
-    if (!err) {
-      const data = parseCoordTxt();
-      fs.open(recordJSON, "r", (err) => {
-        if (!err) {
-          const jsonDAT = require(recordJSON);
+  const jsonPath = recordJson();
 
-          data.forEach((d) => jsonDAT.push(d));
-          fs.writeFileSync(recordJSON, JSON.stringify(jsonDAT, null, 2));
-        } else {
-          fs.writeFileSync(recordJSON, JSON.stringify(data, null, 2));
-        }
-        fs.unlinkSync(recordText);
-      });
-    }
-  });
+  fs.writeFileSync(jsonPath, JSON.stringify(data, null, 2));
+  fs.unlinkSync(recordText);
 }
